backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
email-validator

